Fixed up code to compile under VC++ 8.0.

+++ BalSheetView.cpp
@@ -383,5 +383,5 @@ int CBalSheetView::CompareRows(CRow& oRow1, CRow& oRow2)
 	time_t tValue1 = oRow1[CBalSheet::DATE];
 	time_t tValue2 = oRow2[CBalSheet::DATE];
 
-	return (tValue1 - tValue2);
+	return static_cast<int>(tValue1 - tValue2);
 }
