+++ FixturesView.cpp
@@ -311,5 +311,5 @@ int CFixturesView::CompareRows(CRow& oRow1, CRow& oRow2)
 	time_t tValue1 = oRow1[CFixtures::DATE];
 	time_t tValue2 = oRow2[CFixtures::DATE];
 
-	return (tValue1 - tValue2);
+	return static_cast<int>(tValue1 - tValue2);
 }
